Add files via upload
I created an income-tracker version that does the same thing as expense tracker. Just for income. Next steps are to create exports of the functions and combine the trackers into a budget tracker that will do the math on expenses and income and help users track their left over cash flow.